+++ modules/firehose/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### version / full date
 * [Update/10-Aug-2023] add buffer and cache configurations to fix firehose lag
+* [Update/17-Aug-2023] fix on examples, add custom names for resources